[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -2,13 +2,13 @@
 using namespace std;
 
 int main() {
-	int A, B;
+	long long A, B;
 
 	cin >> A >> B;
 	
 	if (A > B) {
 		cout << A - B - 1 << '\n';
-		for (int i = B + 1; i < A; i++) {
+		for (long long i = B + 1; i < A; i++) {
 			cout << i << ' ';
 		}
 	}
@@ -19,7 +19,7 @@ int main() {
 
 	else {
 		cout << B - A - 1 << '\n';
-		for (int i = A + 1; i < B; i++) {
+		for (long long i = A + 1; i < B; i++) {
 			cout << i << ' ';
 		}
 	}
